Move `appsignal` outside of `javascript/controllers`
