+++ app/javascript/appsignal.js
